Get specialFlow out of plugin base
