+++ plugin-base/src/simple-chat/components/chat-log-section.js
@@ -22,12 +22,7 @@ const ChatLogSection = compose(
   withState('hide', 'setHide', false),
   withState('animate', 'setAnimate', false),
   withHandlers({
-    onClick: ({ onClick, assessmentOptions, setHide, setHideAll }) => ({ type, item }) => {
-      if (type === 'assessmentOption') {
-        assessmentOptions.goToNextStep(item)
-        if (!assessmentOptions.step.multiple) setHideAll(true)
-        return
-      }
+    onClick: ({ onClick, setHide }) => ({ type, item }) => {
       onClick({ type, item })
       setHide(true)
     },
